Moved all rest related model classes under rest and into proper packages.
