+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/controller/EligibilityController.java
@@ -20,16 +20,16 @@ import ca.bc.gov.hlth.hnweb.converter.hl7v2.R15Converter;
 import ca.bc.gov.hlth.hnweb.converter.rapid.RPBSPPL0Converter;
 import ca.bc.gov.hlth.hnweb.converter.rapid.RPBSPPE0Converter;
 import ca.bc.gov.hlth.hnweb.exception.HNWebException;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckEligibilityRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckEligibilityResponse;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckMspCoverageStatusRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckMspCoverageStatusResponse;
-import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnResponse;
-import ca.bc.gov.hlth.hnweb.model.eligibility.LookupPhnRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.LookupPhnResponse;
 import ca.bc.gov.hlth.hnweb.model.rapid.RPBSPPE0;
 import ca.bc.gov.hlth.hnweb.model.rapid.RPBSPPL0;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckEligibilityRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckEligibilityResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckMspCoverageStatusRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckMspCoverageStatusResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.InquirePhnRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.InquirePhnResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.LookupPhnRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.LookupPhnResponse;
 import ca.bc.gov.hlth.hnweb.model.v2.message.E45;
 import ca.bc.gov.hlth.hnweb.model.v2.message.R15;
 import ca.bc.gov.hlth.hnweb.service.EligibilityService;
