+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/controller/EnrollmentController.java
@@ -18,12 +18,12 @@ import ca.bc.gov.hlth.hnweb.converter.hl7v2.R50Converter;
 import ca.bc.gov.hlth.hnweb.converter.hl7v3.FindCandidatesConverter;
 import ca.bc.gov.hlth.hnweb.converter.hl7v3.GetDemographicsConverter;
 import ca.bc.gov.hlth.hnweb.exception.HNWebException;
-import ca.bc.gov.hlth.hnweb.model.EnrollSubscriberRequest;
-import ca.bc.gov.hlth.hnweb.model.EnrollSubscriberResponse;
-import ca.bc.gov.hlth.hnweb.model.NameSearchRequest;
-import ca.bc.gov.hlth.hnweb.model.NameSearchResponse;
-import ca.bc.gov.hlth.hnweb.model.GetPersonDetailsRequest;
-import ca.bc.gov.hlth.hnweb.model.GetPersonDetailsResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.EnrollSubscriberRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.EnrollSubscriberResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.GetPersonDetailsRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.GetPersonDetailsResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.NameSearchRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.NameSearchResponse;
 import ca.bc.gov.hlth.hnweb.model.v2.message.R50;
 import ca.bc.gov.hlth.hnweb.model.v3.FindCandidatesRequest;
 import ca.bc.gov.hlth.hnweb.model.v3.FindCandidatesResponse;
