+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/controller/GroupMemberController.java
@@ -21,9 +21,9 @@ import org.springframework.web.server.ResponseStatusException;
 import ca.bc.gov.hlth.hnweb.converter.rapid.RPBSPED0Converter;
 import ca.bc.gov.hlth.hnweb.converter.rapid.RPBSPEE0Converter;
 import ca.bc.gov.hlth.hnweb.exception.HNWebException;
-import ca.bc.gov.hlth.hnweb.model.StatusEnum;
 import ca.bc.gov.hlth.hnweb.model.rapid.RPBSPED0;
 import ca.bc.gov.hlth.hnweb.model.rapid.RPBSPEE0;
+import ca.bc.gov.hlth.hnweb.model.rest.StatusEnum;
 import ca.bc.gov.hlth.hnweb.model.rest.groupmember.UpdateNumberAndDeptRequest;
 import ca.bc.gov.hlth.hnweb.model.rest.groupmember.UpdateNumberAndDeptResponse;
 import ca.bc.gov.hlth.hnweb.service.GroupMemberService;
