+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/converter/rapid/BaseRapidConverter.java
@@ -4,9 +4,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ca.bc.gov.hlth.hnweb.model.BaseResponse;
-import ca.bc.gov.hlth.hnweb.model.StatusEnum;
 import ca.bc.gov.hlth.hnweb.model.rapid.RPBSHeader;
+import ca.bc.gov.hlth.hnweb.model.rest.BaseResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.StatusEnum;
 import ca.bc.gov.hlth.hnweb.security.SecurityUtil;
 import ca.bc.gov.hlth.hnweb.security.UserInfo;
