+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/rest/eligibility/InquirePhnRequest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnweb.model.eligibility;
+package ca.bc.gov.hlth.hnweb.model.rest.eligibility;
 
 import java.util.ArrayList;
 import java.util.List;
