+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/rest/eligibility/LookupPhnBeneficiary.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnweb.model.eligibility;
+package ca.bc.gov.hlth.hnweb.model.rest.eligibility;
 
 public class LookupPhnBeneficiary {
 	private String phn;
