+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/rest/eligibility/LookupPhnRequest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnweb.model.eligibility;
+package ca.bc.gov.hlth.hnweb.model.rest.eligibility;
 
 public class LookupPhnRequest {
 	private String contractNumber;
