+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/rest/eligibility/LookupPhnResponse.java
@@ -1,9 +1,9 @@
-package ca.bc.gov.hlth.hnweb.model.eligibility;
+package ca.bc.gov.hlth.hnweb.model.rest.eligibility;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import ca.bc.gov.hlth.hnweb.model.BaseResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.BaseResponse;
 
 public class LookupPhnResponse extends BaseResponse {
 	private List<LookupPhnBeneficiary> beneficiaries = new ArrayList<LookupPhnBeneficiary>();
