+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/rest/enrollment/GetPersonDetailsRequest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnweb.model;
+package ca.bc.gov.hlth.hnweb.model.rest.enrollment;
 
 public class GetPersonDetailsRequest {
