+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/rest/enrollment/NameSearchResult.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnweb.model;
+package ca.bc.gov.hlth.hnweb.model.rest.enrollment;
 
 import java.math.BigDecimal;
