+++ backend/src/test/java/ca/bc/gov/hlth/hnweb/controller/EligibilityControllerTest.java
@@ -22,17 +22,17 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 
-import ca.bc.gov.hlth.hnweb.model.StatusEnum;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckEligibilityRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckEligibilityResponse;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckMspCoverageStatusRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.CheckMspCoverageStatusResponse;
-import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnBeneficiary;
-import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnResponse;
-import ca.bc.gov.hlth.hnweb.model.eligibility.LookupPhnBeneficiary;
-import ca.bc.gov.hlth.hnweb.model.eligibility.LookupPhnRequest;
-import ca.bc.gov.hlth.hnweb.model.eligibility.LookupPhnResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.StatusEnum;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckEligibilityRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckEligibilityResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckMspCoverageStatusRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.CheckMspCoverageStatusResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.InquirePhnBeneficiary;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.InquirePhnRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.InquirePhnResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.LookupPhnBeneficiary;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.LookupPhnRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.eligibility.LookupPhnResponse;
 import ca.bc.gov.hlth.hnweb.security.SecurityUtil;
 import ca.bc.gov.hlth.hnweb.security.UserInfo;
 import ca.bc.gov.hlth.hnweb.util.V2MessageUtil;
