+++ backend/src/test/java/ca/bc/gov/hlth/hnweb/controller/EnrollmentControllerTest.java
@@ -20,13 +20,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 
-import ca.bc.gov.hlth.hnweb.model.EnrollSubscriberRequest;
-import ca.bc.gov.hlth.hnweb.model.EnrollSubscriberResponse;
-import ca.bc.gov.hlth.hnweb.model.NameSearchRequest;
-import ca.bc.gov.hlth.hnweb.model.NameSearchResponse;
-import ca.bc.gov.hlth.hnweb.model.GetPersonDetailsRequest;
-import ca.bc.gov.hlth.hnweb.model.GetPersonDetailsResponse;
-import ca.bc.gov.hlth.hnweb.model.StatusEnum;
+import ca.bc.gov.hlth.hnweb.model.rest.StatusEnum;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.EnrollSubscriberRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.EnrollSubscriberResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.GetPersonDetailsRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.GetPersonDetailsResponse;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.NameSearchRequest;
+import ca.bc.gov.hlth.hnweb.model.rest.enrollment.NameSearchResponse;
 import ca.bc.gov.hlth.hnweb.security.SecurityUtil;
 import ca.bc.gov.hlth.hnweb.security.UserInfo;
 import ca.bc.gov.hlth.hnweb.utils.TestUtil;
